requirements.txt
ultralytics>=8.0.0
pyyaml>=6.0
opencv-python>=4.6.0
Pillow>=9.0.0
shapely>=2.0.0
neo4j>=5.0.0
numpy>=1.21.0
docker>=6.0.0
pytest>=7.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0